refactor(Progress): rename directory
