+++ components/Article/Progress/ProgressPrompt.js
